VolunteersProfile/backend/requirements.txt
asgiref==3.7.2
astroid==2.15.5
dill==0.3.6
Django==4.2.2
django-cors-headers==4.2.0
django-dotenv==1.4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==10.0.0
platformdirs==3.7.0
psycopg2-binary==2.9.6
PyJWT==1.7.1
pylint==2.17.4
pytz==2023.3
sqlparse==0.4.4
tomlkit==0.11.8
wrapt==1.15.0
